server/requirements.txt
flask
flask_cors
psycopg2-binary
pycryptodome
